Fix -Wsign-compare: heap overflow in cc_xml_load_file() on ftell() error
cc_xml_load_file() never checked ftell()'s return value, which is -1
on error -- notably whenever the stream doesn't support seeking, e.g.
a FIFO/pipe. On that path, bufsize (a signed long) became -1, `new
char[bufsize + 1]` allocated a ZERO-byte buffer, and the
signed/unsigned mismatch flagged by -Wsign-compare meant `bufsize`
converted to SIZE_MAX in both `pos != bufsize` and `bufsize - pos` --
the latter being fread()'s size argument, so fread() was asked to
read up to SIZE_MAX bytes into that 0-byte buffer.

Confirmed with a reproducer using a real POSIX named pipe (an
ordinary, realistic way for ftell() to fail): glibc's _FORTIFY_SOURCE
catches it immediately with "*** buffer overflow detected ***"
(SIGABRT), and a gdb backtrace confirms fread() was called with
n=18446744073709551615 (SIZE_MAX) at utils.cpp:54, exactly as
predicted.

Fix: check ftell()'s return value for < 0 and fail cleanly (close the
file, return NULL, matching the existing !fd convention) before using
it, then store it as size_t for the rest of the function -- this
directly resolves the sign mismatch since there is no longer any
signed/negative value in the computation. Also added a standalone
POSIX reproducer (testsuite/reproducers/xml-load-file-ftell-error/,
not run by default -- needs fork()/mkfifo(), not available on all
platforms this test-suite supports) and a portable embedded regression
test (load_file_roundtrip in src/xml/utils.cpp) covering the ordinary,
seekable-file happy path.

Also fixed SoInput_FileInfo::readUnsignedIntegerString()'s
`readString.length() - oldLength < minSize` (size_t vs int): audited
this one and found no live bug -- within this function readString is
only ever appended to (readDigits()/readHexDigits() both just do
readString += c in a loop), so the subtraction can never underflow.
Changed minSize from int to size_t since it's always a small
non-negative size threshold (1 or 3); this is a type-hygiene fix with
no behavior change.

Full clean rebuild: 0 errors, 0 -Wsign-compare remaining (was 2), total
warnings 859->857. Full CoinTests suite (327 tests) passes under a
plain build and under ASan+UBSan, with no new sanitizer findings, and
cc_xml_load_file()'s ordinary regular-file behavior verified unchanged
(exact content round-trip).

+++ src/io/SoInput_FileInfo.cpp
@@ -494,7 +494,7 @@ SbBool
 SoInput_FileInfo::readUnsignedIntegerString()
 {
   assert(!this->isBinary());
-  int minSize = 1;
+  size_t minSize = 1;
   char c;
   size_t oldLength = readString.length();
